Fixed some log show bug
