+++ src/station/src/selftest_lora.c
@@ -44,25 +44,25 @@ void selftest_lora () {
     s2e_joineuiFilter = joineuiFilter;
 
     T = "\x00_______________";  // too short
-    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)T, 1, NULL));
+    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)T, 1, NULL, NULL));
     T = "\x03_______________";  // bad major version
-    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)T, 16, NULL));
+    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)T, 16, NULL, NULL));
 
     B.pos = 0;
     T = "\x20_______________";  // join accept
-    TCHECK(s2e_parse_lora_frame(&B, (const u1_t*)T, 16, NULL));
+    TCHECK(s2e_parse_lora_frame(&B, (const u1_t*)T, 16, NULL, NULL));
     xeos(&B);
     TCHECK(strcmp("\"msgtype\":\"jacc\",\"FRMPayload\":\"205F5F5F5F5F5F5F5F5F5F5F5F5F5F5F\"", B.buf) == 0);
 
     B.pos = 0;
     T = "\xE0_______________";  // proprietary frame
-    TCHECK(s2e_parse_lora_frame(&B, (const u1_t*)T, 16, NULL));
+    TCHECK(s2e_parse_lora_frame(&B, (const u1_t*)T, 16, NULL, NULL));
     xeos(&B);
     TCHECK(strcmp("\"msgtype\":\"propdf\",\"FRMPayload\":\"E05F5F5F5F5F5F5F5F5F5F5F5F5F5F5F\"", B.buf) == 0);
 
     B.pos = 0;
     const char* Tjreq = "\x00\x01\x23\x45\x67\x89\xAB\xCD\xEF\xF1\xE3\xF5\xE7\xF9\xEB\xFD\xEF\xF0\xF1\xA0\xA1\xA2\xA3";  // jreq
-    TCHECK(s2e_parse_lora_frame(&B, (const u1_t*)Tjreq, 23, NULL));
+    TCHECK(s2e_parse_lora_frame(&B, (const u1_t*)Tjreq, 23, NULL, NULL));
     xeos(&B);
     TCHECK(strcmp("\"msgtype\":\"jreq\",\"MHdr\":0,"
                   "\"JoinEui\":\"EF-CD-AB-89-67-45-23-01\","
@@ -70,19 +70,19 @@ void selftest_lora () {
                   "\"DevNonce\":61936,\"MIC\":-1549622880", B.buf) == 0);
     // Too short
     B.pos = 0;
-    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)Tjreq, 22, NULL));
+    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)Tjreq, 22, NULL, NULL));
     // Filter enabled
     B.pos = 0;
     memcpy(s2e_joineuiFilter, euiFilter1, sizeof(euiFilter1));  // jreq removed
-    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)Tjreq, 23, NULL));
+    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)Tjreq, 23, NULL, NULL));
     B.pos = 0;
     memcpy(s2e_joineuiFilter, euiFilter2, sizeof(euiFilter2));  // jreq passes
-    TCHECK(s2e_parse_lora_frame(&B, (const u1_t*)Tjreq, 23, NULL));
+    TCHECK(s2e_parse_lora_frame(&B, (const u1_t*)Tjreq, 23, NULL, NULL));
     s2e_joineuiFilter[0] = 0;
 
     B.pos = 0;
     const char* Tdaup1 = "\x40\xAB\xCD\xEF\xFF\x01\xF3\xF4\xFF\x20\x21\x22\xA0\xA1\xA2\xA3";  // daup
-    TCHECK(s2e_parse_lora_frame(&B, (const u1_t*)Tdaup1, 12+1+3, NULL));
+    TCHECK(s2e_parse_lora_frame(&B, (const u1_t*)Tdaup1, 12+1+3, NULL, NULL));
     xeos(&B);
     TCHECK(strcmp("\"msgtype\":\"updf\","
                   "\"MHdr\":64,\"DevAddr\":-1061461,\"FCtrl\":1,\"FCnt\":62707,"
@@ -90,11 +90,11 @@ void selftest_lora () {
                   "\"MIC\":-1549622880", B.buf) == 0);
     // Too short
     B.pos = 0;
-    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)Tdaup1, 12, NULL));
+    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)Tdaup1, 12, NULL, NULL));
     // Filtered
     B.pos = 0;
     s2e_netidFilter[0] = s2e_netidFilter[1] = s2e_netidFilter[2] = s2e_netidFilter[3] = 0;
-    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)Tdaup1, 12+1+3, NULL));
+    TCHECK(!s2e_parse_lora_frame(&B, (const u1_t*)Tdaup1, 12+1+3, NULL, NULL));
 
     free(jsonbuf);
 }
